Fix site title attribute logic (#121)
The site title attribute had some logic that would change the site title if a project modal was open, but it caused the title to permanently be set to the title of the last entry in the _posts directory. I removed the if/else logic to keep the site title to be what is specified in _config.yml in the "title" field.